feature_extract

+++ mmaction/models/extractors/video_extractor.py
@@ -1,6 +1,8 @@
 import torch.nn as nn
 import torch
 import numpy as np
+import os.path as osp
+import sys
 
 from .. import builder
 from .base import FeatureExtractor
@@ -19,6 +21,19 @@ class VideoExtractor(FeatureExtractor):
 
     def forward_test(self, x, img_metas=None):
         # x.shape = BNCHW
+        print(type(img_metas))
+        sys.exit(0)
+        batch, frames = x.shape[:2]
+        idxes = []
+        for i in range(batch):
+            feature_path = img_metas[i]["featurepath"]
+            if not osp.exists(feature_path):
+                idxes.append(i)
+        idxes = np.array(idxes)
+        if len(idxes) == 0:
+            return
+        x = x[idxes]
+        img_metas = img_metas[idxes]
         with torch.no_grad():
             batch, frames = x.shape[:2]
             x = x.reshape((-1,) + x.shape[2:])  # BN * CHW
@@ -28,7 +43,7 @@ class VideoExtractor(FeatureExtractor):
             for i in range(batch):
                 feature = x[i].unsqueeze(0).cpu().detach().numpy()  # 1, 32, 2048
                 feature_path = img_metas[i]["featurepath"]
-                if feature_path.endswith(".npy"):
+                if feature_path.endswith(".npy") and not osp.exists(feature_path):
                     np.save(feature_path, feature)
 
     def forward(self, imgs, return_loss=False, img_metas=None):
